+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Use scm-code-editor-plugin for syntax highlighting ([#14](https://github.com/scm-manager/scm-script-plugin/pull/14))
+
 ## [2.1.3] - 2020-09-02
 
 ### Fixed